pip_requirements.txt
aiofiles==0.8.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitarray==2.5.1
bleach==5.0.0
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
ci-info==0.2.0
click==8.1.3
configobj==5.0.6
configparser==5.2.0
cryptography==36.0.2
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
debugpy==1.6.0
decorator==5.1.1
deep-translator==1.8.3
deepl-translate==1.2.0
defusedxml==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints==0.4
etelemetry==0.3.0
executing==0.8.3
fastjsonschema==2.15.3
fasttext==0.9.2
fastwer==0.1.3
filelock==3.7.0
fonttools==4.33.3
fpdf==1.7.2
frontend==0.0.3
future==0.18.2
gensim==4.2.0
grobid-client==0.8.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
hyperframe==5.2.0
idna==2.10
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
langcodes==3.3.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
nibabel==3.2.2
nipype==1.8.0
nltk==3.7
notebook==6.4.11
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathy==0.6.1
pdfminer==20191125
pdfminer.six==20220506
pdfreader==0.1.11
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==5.1.3
Pillow==9.1.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
prov==2.0.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.9.2
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.8.2
pydot==1.4.2
pyenchant==3.2.2
Pygments==2.12.0
PyMuPDF==1.19.6
pyparsing==3.0.9
PyPDF4==1.27.0
pyrsistent==0.18.1
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
pyxnat==1.4
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
rdflib==6.1.1
regex==2022.4.24
requests==2.27.1
rfc3986==1.5.0
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.2.0
starlette==0.20.0
terminado==0.13.3
tesseract-ocr==0.0.1
thinc==8.0.16
tika==1.24
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
traits==6.3.2
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
